requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.12.0
httpx>=0.27.0
tenacity>=8.2.0
edge-tts>=7.0.0
